Dropping 3M32 resistor to 1M4.
The DC/DC converter is currently trimmed to 5.1 V. Since the
Raspberry Pi can take more than that, we increase the voltage
to 5.25 V. This appears to be the recommended operating voltage.

In order to achieve this, a trimming resistor of 1M375 is
required. We use 1M4. Since we also use this resistor value to
dissipate stray charge on FET and microcontroller inputs, those
resistors are dropped in value as well. This is a good thing since
the value was uncomfortably high (and still remains somewhat high
even after the change).

Fixes #2.