Adding diode to address case where DC/DC converter is reverse biased
The datasheet of the DC/DC converter recommends a (normally) reverse
biased diode across the regulator. This is to prevent damage in case
the output is at a higher voltage than the input.

Since this condition would apply if somebody powers the Raspberry Pi
directly (i.e. not using the hat), it is strongly recommended to add
the diode.

Fixes #1